fundoo_notes/requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.0
blinker==1.7.0
celery==5.4.0
certifi==2024.2.2
cffi==1.17.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
cryptography==43.0.0
customtkinter==5.2.2
darkdetect==0.8.0
distlib==0.3.8
Django==5.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.15.4
Flask==1.1.4
Flask-SQLAlchemy==2.5.1
greenlet==3.0.3
idna==2.10
importlib_resources==6.4.4
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==1.1.0
Jinja2==3.0.1
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
loguru==0.7.2
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasynchat==1.0.4
pyasyncore==1.0.4
pycparser==2.22
PyJWT==2.9.0
pypiwin32==223
pytest==8.2.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
requests==2.25.1
rpds-py==0.20.0
scikit-learn==1.4.1.post1
scipy==1.13.0
setuptools==69.2.0
six==1.16.0
SQLAlchemy==2.0.29
sqlparse==0.5.1
swagger-spec-validator==3.0.4
threadpoolctl==3.4.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
Werkzeug==1.0.1
win10toast==0.9
win32-setctime==1.1.0
